Extract NameofRules from TransformWorkerProgram into its own file
Un-nest the helper so later pipeline splits can call it without a
Program nested-type qualifier. Nested-type members are dedented to
compilation-unit scope; call sites now use NameofRules directly.
